# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API service to store and modify the product details description of any goods",
        "title": "Details API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/details/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves all product details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/details.Details"
                            }
                        }
                    }
                }
            }
        },
        "/details/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "summary": "Creates a new product-details from the received json document",
                "parameters": [
                    {
                        "description": "product-details schema",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/details.DetailsWithoutId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/details.Details"
                        }
                    }
                }
            }
        },
        "/details/one/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves product-details on given ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product details ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/details.Details"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "details.Details": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "product-name": {
                    "type": "string"
                }
            }
        },
        "details.DetailsWithoutId": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "product-name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}